requirements.txt
altgraph==0.17.2
pycryptodomex==3.15.0
pyinstaller>5.3
pyinstaller-hooks-contrib==2022.10
pyperclip==1.8.2
